+++ CHANGELOG.md
@@ -171,6 +171,7 @@ See <https://commonmark.thephpleague.com/2.0/upgrading/> for detailed informatio
 
  - Removed support for PHP 7.1
  - Removed all previously-deprecated functionality:
+   - Removed the ability to pass custom `Environment` instances into the `CommonMarkConverter` and `GithubFlavoredMarkdownConverter` constructors
    - Removed the `Converter` class and `ConverterInterface`
    - Removed the `bin/commonmark` script
    - Removed the `Html5Entities` utility class